code-examples/chapter-11/standalone/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

